% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTD_gname.R
\name{GTD_gname}
\alias{GTD_gname}
\title{Generates a GTD gname variable from a variable that contains the names of terrorist groups.
\code{GTD_gname} generates a GTD gname variable using alternative names and spellings of group names.}
\usage{
GTD_gname(data, gname)
}
\arguments{
\item{data}{data.frame}

\item{gname}{The name of the variable in the data that contains the names of terrorist groups.}
}
\value{
the output will be a vector or the same length as gname.
}
\description{
Generates a GTD gname variable from a variable that contains the names of terrorist groups.
\code{GTD_gname} generates a GTD gname variable using alternative names and spellings of group names.
}
